Use async-signals for controller traps.
