+++ async-container.gemspec
@@ -25,4 +25,5 @@ Gem::Specification.new do |spec|
 	spec.required_ruby_version = ">= 3.3"
 	
 	spec.add_dependency "async", "~> 2.22"
+	spec.add_dependency "async-signals", "~> 0.1"
 end
