+++ lib/async/container.rb
@@ -4,7 +4,6 @@
 # Copyright, 2017-2025, by Samuel Williams.
 
 require_relative "container/events"
-require_relative "container/signals"
 require_relative "container/controller"
 
 # @namespace
